Indeed it is true I get more done in pure silence
When I lack said pure silence, I do become a tad frustrated. But when I do get the opportunity for mostly pure silence, why do I opt to fill it with meaningless noise? It is only depriving me the opportunity to create my life's richest work, increase my wealth, and maintain a level of focus, level headedness, and empathy that is otherwise not possible when hearing and seeing the negativities and decractors of the world.